Añadir ejecutable benchmark y establecer directorio de salida en CMakeLists.txt; mejorar formato y legibilidad en io.cpp.

+++ utils/io.cpp
@@ -1,4 +1,4 @@
-//Funciones comunes (lectura, timer, estadísticas)
+// Funciones comunes (lectura, timer, estadísticas)
 
 #include "io.h"
 #include <fstream>
@@ -9,13 +9,17 @@
 
 namespace fs = std::filesystem;
 
-std::string leerDocumentosDesdeCarpeta(const std::string& carpeta, std::vector<std::string>& nombresDoc, std::vector<int>& cortes) {
+std::string leerDocumentosDesdeCarpeta(const std::string &carpeta, std::vector<std::string> &nombresDoc, std::vector<int> &cortes)
+{
     std::vector<std::pair<std::string, std::string>> archivosContenido;
 
-    for (const auto& entry : fs::directory_iterator(carpeta)) {
-        if (entry.path().extension() == ".txt") {
+    for (const auto &entry : fs::directory_iterator(carpeta))
+    {
+        if (entry.path().extension() == ".txt")
+        {
             std::ifstream file(entry.path());
-            if (!file.is_open()) continue;
+            if (!file.is_open())
+                continue;
 
             std::string contenido((std::istreambuf_iterator<char>(file)), std::istreambuf_iterator<char>());
             std::string nombreArchivo = entry.path().filename().string();
@@ -31,7 +35,8 @@ std::string leerDocumentosDesdeCarpeta(const std::string& carpeta, std::vector<s
     nombresDoc.clear();
     cortes.clear();
 
-    for (const auto& [nombre, contenido] : archivosContenido) {
+    for (const auto &[nombre, contenido] : archivosContenido)
+    {
         textoTotal += contenido + '$';
         acumulado = textoTotal.size();
         cortes.push_back(acumulado);
@@ -41,16 +46,19 @@ std::string leerDocumentosDesdeCarpeta(const std::string& carpeta, std::vector<s
     return textoTotal;
 }
 
-std::string concatenarDocumentosSeleccionados(const std::string& carpeta,
-                                             const std::vector<std::string>& docsSeleccionados,
-                                             std::vector<int>& cortes) {
+std::string concatenarDocumentosSeleccionados(const std::string &carpeta,
+                                              const std::vector<std::string> &docsSeleccionados,
+                                              std::vector<int> &cortes)
+{
     std::string textoTotal;
     int acumulado = 0;
     cortes.clear();
 
-    for (const auto& nombreArchivo : docsSeleccionados) {
+    for (const auto &nombreArchivo : docsSeleccionados)
+    {
         std::ifstream file(carpeta + "/" + nombreArchivo);
-        if (!file.is_open()) continue;
+        if (!file.is_open())
+            continue;
 
         std::string contenido((std::istreambuf_iterator<char>(file)), std::istreambuf_iterator<char>());
         textoTotal += contenido + '$';
@@ -60,9 +68,12 @@ std::string concatenarDocumentosSeleccionados(const std::string& carpeta,
     return textoTotal;
 }
 
-int obtenerDocumento(int pos, const std::vector<int>& cortes) {
-    for (int i = 0; i < cortes.size(); ++i) {
-        if (pos < cortes[i]) {
+int obtenerDocumento(int pos, const std::vector<int> &cortes)
+{
+    for (size_t i = 0; i < cortes.size(); ++i)
+    {
+        if (pos < cortes[i])
+        {
             return i + 1;
         }
     }
